+++ test/integration/IceIntegrationTest.cpp
@@ -1,0 +1,20 @@
+#include "test/integration/IntegrationTest.h"
+
+/*
+    FakeStunServer stunServer(transport::SocketAddress::parse("64.233.165.127", 19302), internet);
+    fakenet::Firewall firewallLocal(transport::SocketAddress::parse("216.93.246.10", 0), internet);
+    fakenet::Firewall firewallVpn(transport::SocketAddress::parse("2001:1ba8:148c:bf00:9181:9fb2:8def:a66e", 0),
+        internet);
+
+    fakenet::FakeEndpoint endpoint1(transport::SocketAddress::parse("192.168.1.10", 2000), firewallLocal);
+    fakenet::FakeEndpoint endpointTcp1(transport::SocketAddress::parse("192.168.1.10", 6700), firewallLocal);
+    fakenet::FakeEndpoint endpoint2(transport::SocketAddress::parse("fc00:1808:1808:1808:1808:1808:1808:1100", 3000),
+        firewallVpn);
+
+    fakenet::FakeEndpoint smbEndpoint(transport::SocketAddress::parse("111.11.1.11", 10000), internet);
+    fakenet::FakeEndpoint smbEndpointIpv6(transport::SocketAddress::parse("2600:1900:4080:1627:0:26:0:0", 10000),
+        internet);
+    fakenet::FakeEndpoint smbTcpServerEndpoint(transport::SocketAddress::parse("111.11.1.11", 4443),
+        internet,
+        ice::TransportType::TCP);
+*/
