+++ test/integration/emulator/SfuClient.h
@@ -3,8 +3,11 @@
 #include "AudioSource.h"
 #include "FakeVideoDecoder.h"
 #include "api/SimulcastGroup.h"
+#include "bridge/RtpMap.h"
 #include "bridge/engine/PacketCache.h"
+#include "bridge/engine/SsrcInboundContext.h"
 #include "bridge/engine/VideoMissingPacketsTracker.h"
+#include "codec/Opus.h"
 #include "codec/Vp8Header.h"
 #include "legacyapi/DataChannelMessage.h"
 #include "memory/Array.h"
@@ -12,12 +15,14 @@
 #include "memory/PacketPoolAllocator.h"
 #include "rtp/RtcpFeedback.h"
 #include "rtp/RtcpHeader.h"
+#include "test/integration/emulator/ApiChannel.h"
 #include "test/transport/FakeNetwork.h"
 #include "transport/DataReceiver.h"
 #include "transport/RtcTransport.h"
 #include "transport/TransportFactory.h"
 #include "transport/dtls/SslDtls.h"
 #include "utils/IdGenerator.h"
+#include "utils/Pacer.h"
 #include "utils/Span.h"
 #include "utils/StringBuilder.h"
 #include "webrtc/WebRtcDataStream.h"
